Day 23: Both parts
Unfortunately, I was at Star Wars when this puzzle came out, so no
leaderboard for me :(.

Part 1 is a straightforward VM implementation (mostly copiable from Day
18).

Part 2 was much more interesting.  I had to reverse engineer the
innermost two loops before this ran quickly enough to complete.  Maybe
in a faster language like C you would only have to reverse one loop.  Or
none?  I'll give that a shot in a second.

Anyway, the general structure of the program is this:

In part 2, it counts the number of factorable numbers in the range
[107900 (my puzzle input), 107900 + 17000 (probably 2017 reference)]
inclusive, with a step size of 17.  (So, 1000 primality tests.)

The implementation of primality testing is extremely naive; it is
something like this (Python):

def isprime(b):
  prime = True
  for d in xrange(2, b):
    for e in xrange(2, b):
      if b == d * e:
        prime = False
  return prime